dev-requirements.txt
-e .
ipdb==0.11
pytest==4.2.0
pytest-cov==2.6.1
pre-commit==1.10.5
pyhamcrest==1.9.0
pytest-localserver==0.5.0
attrdict==2.0.0
